+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/converters/urdf_converter.py
@@ -6,9 +6,13 @@
 
 from __future__ import annotations
 
+import os
+
 import omni.kit.commands
+import omni.usd
 from omni.isaac.core.utils.extensions import enable_extension
 from omni.isaac.version import get_version
+from pxr import Usd
 
 from .asset_converter_base import AssetConverterBase
 from .urdf_converter_cfg import UrdfConverterCfg
@@ -81,6 +85,14 @@ class UrdfConverter(AssetConverterBase):
             import_config=import_config,
             dest_path=self.usd_path,
         )
+        # fix the issue that material paths are not relative
+        if self.cfg.make_instanceable:
+            usd_path = os.path.join(self.usd_dir, self.usd_instanceable_meshes_path)
+            stage = Usd.Stage.Open(usd_path)
+            # resolve all paths relative to layer path
+            source_layer = stage.GetRootLayer()
+            omni.usd.resolve_paths(source_layer.identifier, source_layer.identifier)
+            stage.Save()
 
     """
     Helper methods.